+++ src/main/java/example/infrastructure/datasource/worker/WorkerMapper.java
@@ -17,19 +17,19 @@ public interface WorkerMapper {
 
     void deleteWorkerName(@Param("workerNumber") WorkerNumber workerNumber);
 
-    void insertWorkerNameHistory(@Param("id") Long id, @Param("workerNumber") WorkerNumber workerNumber, @Param("name") Name name);
+    void insertWorkerNameHistory(@Param("id") Integer id, @Param("workerNumber") WorkerNumber workerNumber, @Param("name") Name name);
 
     void insertWorkerName(@Param("workerNumber") WorkerNumber workerNumber, @Param("name") Name workerName);
 
     void deleteWorkerPhoneNumber(@Param("workerNumber") WorkerNumber workerNumber);
 
-    void insertWorkerPhoneNumberHistory(@Param("id") Long id, @Param("workerNumber") WorkerNumber workerNumber, @Param("phoneNumber") PhoneNumber phoneNumber);
+    void insertWorkerPhoneNumberHistory(@Param("id") Integer id, @Param("workerNumber") WorkerNumber workerNumber, @Param("phoneNumber") PhoneNumber phoneNumber);
 
     void insertWorkerPhoneNumber(@Param("workerNumber") WorkerNumber workerNumber, @Param("phoneNumber") PhoneNumber phoneNumber);
 
     void deleteWorkerMailAddress(@Param("workerNumber") WorkerNumber workerNumber);
 
-    void insertWorkerMailAddressHistory(@Param("id") Long id, @Param("workerNumber") WorkerNumber workerNumber, @Param("mailAddress") MailAddress mailAddress);
+    void insertWorkerMailAddressHistory(@Param("id") Integer id, @Param("workerNumber") WorkerNumber workerNumber, @Param("mailAddress") MailAddress mailAddress);
 
     void insertWorkerMailAddress(@Param("workerNumber") WorkerNumber workerNumber, @Param("mailAddress") MailAddress mailAddress);
 
@@ -39,11 +39,11 @@ public interface WorkerMapper {
 
     void insertExpireContract(@Param("workerNumber") WorkerNumber workerNumber);
 
-    long newWorkerNumber();
+    Integer newWorkerNumber();
 
-    long newWorkerNameIdentifier();
+    Integer newWorkerNameIdentifier();
 
-    long newWorkerPhoneNumberIdentifier();
+    Integer newWorkerPhoneNumberIdentifier();
 
-    long newWorkerMailAddressIdentifier();
+    Integer newWorkerMailAddressIdentifier();
 }
